# Faster parsing by changing lhs of assignment-expression to conditional-expression.