@vickenty 2020-05-25
Thanks, this is great! 

I think this is a reasonable compromise to make, given the huge difference in performance. It shouldn't make user's life much harder, since not all unary expressions are assignable, and lvalueness needs to be checked anyway.

1. Is it possible to make benchmark smaller? Adding 10M to the repo just for this seems excessive.
1. Please also add a comment to the grammar file explaining why we deviate from the standard grammar.